wallet-instance encoder stuff in io-wallet-user-func
