+++ apps/io-wallet-user-func/src/infra/http/handlers/get-current-wallet-instance-status.ts
@@ -1,9 +1,9 @@
+import { WalletInstanceToStatus } from "@/encoders/wallet-instance";
 import { getCurrentWalletInstance } from "@/wallet-instance";
 import * as H from "@pagopa/handler-kit";
 import * as RTE from "fp-ts/lib/ReaderTaskEither";
 import { pipe } from "fp-ts/lib/function";
 import { logErrorAndReturnResponse } from "io-wallet-common/infra/http/error";
-import { WalletInstanceToStatus } from "io-wallet-common/wallet-instance";
 
 import { requireWhitelistedUserFromToken } from "../whitelisted-user";
