+++ packages/io-wallet-common/src/wallet-instance.ts
@@ -42,20 +42,3 @@ export const WalletInstance = t.union([
 ]);
 
 export type WalletInstance = t.TypeOf<typeof WalletInstance>;
-
-const WalletInstanceStatus = t.type({
-  id: NonEmptyString,
-  is_revoked: t.boolean,
-});
-
-type WalletInstanceStatus = t.TypeOf<typeof WalletInstanceStatus>;
-
-export const WalletInstanceToStatus: t.Encoder<
-  WalletInstance,
-  WalletInstanceStatus
-> = {
-  encode: ({ id, isRevoked }) => ({
-    id,
-    is_revoked: isRevoked,
-  }),
-};
